+++ src/components/AdminLanguage/AdminLanguage.jsx
@@ -16,6 +16,10 @@ import {
   TablePagination,
 } from "@mui/material";
 
+//components
+import AdminLanguageRow from "./AdminLanguageRow";
+
+
 function AdminLanguage() {
   const dispatch = useDispatch();
   const [page, setPage] = useState(0);
@@ -26,13 +30,6 @@ function AdminLanguage() {
     (store) => store.adminReducer.adminLanguagesReducer
   );
 
-  // Defining what to do when edit or delete are pressed:
-  const handleEdit = () => {
-    console.log("Edit");
-  };
-  const handleDelete = () => {
-    console.log("Delete");
-  };
 
   // The below 2 functions allow there to be multiple pages on the table.
   const handleChangePage = (event, newPage) => {
@@ -64,33 +61,15 @@ function AdminLanguage() {
                 <TableCell >Edit/Delete</TableCell>
               </TableRow>
             </TableHead>
-            <TableBody>
+            
               {languages
                 .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                 .map((language) => {
                   return (
-                    <TableRow hover role="checkbox" tabIndex={-1}>
-                      <TableCell>{language.language}</TableCell>
-                      <TableCell>{language.glottocode}</TableCell>
-                      <TableCell>{language.global_speakers}</TableCell>
-                      <TableCell>{language.sc_speakers}</TableCell>
-                      <TableCell>{language.endonym}</TableCell>
-                      <TableCell>
-                        <Button
-                          sx={{ mr: 1 }}
-                          variant="contained"
-                          onClick={handleEdit}
-                        >
-                          Edit
-                        </Button>
-                        <Button variant="contained" onClick={handleDelete}>
-                          Delete
-                        </Button>
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-            </TableBody>
+                    <TableBody key={language.id}>
+                    <AdminLanguageRow language={language} />
+                    </TableBody>
+                  );})}
           </Table>
         </TableContainer>
         <TablePagination
